transferindo todas as divisões do home para a pasta Widgets
